fix: Run setup only once

+++ ci-scripts/load-test.sh
@@ -32,39 +32,42 @@ elif [ "$TEST_RUN" == "./run-build-image.yaml" ]; then
         debug "Skipping initial config as namespace utils already exists"
     fi
 elif [ "$TEST_RUN" == "./run-image-signing.yaml" ]; then
-    [ -d push-fake-image ] || git clone https://github.com/jhutar/push-fake-image.git
-    kubectl apply -f push-fake-image/pipeline.yaml
-    # Configure Chains as per https://tekton.dev/docs/chains/signed-provenance-tutorial/#configuring-tekton-chains
-    kubectl patch TektonConfig/config --type='merge' -p='{"spec":{"chain":{"artifacts.taskrun.format": "slsa/v1"}}}'
-    kubectl patch TektonConfig/config --type='merge' -p='{"spec":{"chain":{"artifacts.taskrun.storage": "oci"}}}'
-    kubectl patch TektonConfig/config --type='merge' -p='{"spec":{"chain":{"artifacts.oci.storage": "oci"}}}'
-    kubectl patch TektonConfig/config --type='merge' -p='{"spec":{"chain":{"transparency.enabled": "false"}}}'   # this is the only difference from the docs
-    export COSIGN_PASSWORD=reset
-    before=$( date +%s )
-    while ! cosign generate-key-pair -d k8s://openshift-pipelines/signing-secrets; do
-        now=$( date +%s )
-        [ $(( $now - $before )) -gt 300 ] && fatal "Was not able to create signing-secrets secret in time"
-        debug "Waiting for next attempt for creation of signing-secrets secret"
-        oc -n openshift-pipelines get secrets || true   # FIXME
-        oc -n openshift-pipelines describe secrets/signing-secrets || true   # FIXME
-        oc -n openshift-pipelines get secrets/signing-secrets -o yaml || true   # FIXME
-        oc -n openshift-pipelines delete secrets/signing-secrets || true   # FIXME
-        sleep 10
-    done
-    wait_for_entity_by_selector 300 openshift-pipelines deployment app.kubernetes.io/name=controller,app.kubernetes.io/part-of=tekton-chains
-    oc -n openshift-pipelines rollout restart deployment/tekton-chains-controller
-    oc -n openshift-pipelines rollout status deployment/tekton-chains-controller
-    oc -n openshift-pipelines wait --for=condition=ready --timeout=300s pod -l app.kubernetes.io/part-of=tekton-chains
-    # ImageStreamTag to push to
-    oc -n benchmark create imagestream test
-    # SA to talk to internal registry
-    oc -n benchmark create serviceaccount perf-test-registry-sa
-    oc policy add-role-to-user registry-viewer system:serviceaccount:benchmark:perf-test-registry-sa   # pull
-    oc policy add-role-to-user registry-editor system:serviceaccount:benchmark:perf-test-registry-sa   # push
-    # Customized PipelineRun
-    dockerconfig_secret_name=$( oc -n benchmark get serviceaccount perf-test-registry-sa -o json | jq --raw-output '.imagePullSecrets[0].name' )
-    TEST_RUN="./push-fake-image/run-modified.yaml"
-    cat ./push-fake-image/run.yaml | sed "s/DOCKERCONFIG_SECRET_NAME/$dockerconfig_secret_name/g" >$TEST_RUN
+    if ! [ -d push-fake-image ]; then
+        git clone https://github.com/jhutar/push-fake-image.git
+        kubectl apply -f push-fake-image/pipeline.yaml
+        # Configure Chains as per https://tekton.dev/docs/chains/signed-provenance-tutorial/#configuring-tekton-chains
+        kubectl patch TektonConfig/config --type='merge' -p='{"spec":{"chain":{"artifacts.taskrun.format": "slsa/v1"}}}'
+        kubectl patch TektonConfig/config --type='merge' -p='{"spec":{"chain":{"artifacts.taskrun.storage": "oci"}}}'
+        kubectl patch TektonConfig/config --type='merge' -p='{"spec":{"chain":{"artifacts.oci.storage": "oci"}}}'
+        kubectl patch TektonConfig/config --type='merge' -p='{"spec":{"chain":{"transparency.enabled": "false"}}}'   # this is the only difference from the docs
+        export COSIGN_PASSWORD=reset
+        before=$( date +%s )
+        while ! cosign generate-key-pair -d k8s://openshift-pipelines/signing-secrets; do
+            now=$( date +%s )
+            [ $(( $now - $before )) -gt 300 ] && fatal "Was not able to create signing-secrets secret in time"
+            debug "Waiting for next attempt for creation of signing-secrets secret"
+            oc -n openshift-pipelines get secrets || true   # FIXME
+            oc -n openshift-pipelines describe secrets/signing-secrets || true   # FIXME
+            oc -n openshift-pipelines get secrets/signing-secrets -o yaml || true   # FIXME
+            oc -n openshift-pipelines delete secrets/signing-secrets || true   # FIXME
+            sleep 10
+        done
+        wait_for_entity_by_selector 300 openshift-pipelines deployment app.kubernetes.io/name=controller,app.kubernetes.io/part-of=tekton-chains
+        oc -n openshift-pipelines rollout restart deployment/tekton-chains-controller
+        oc -n openshift-pipelines rollout status deployment/tekton-chains-controller
+        oc -n openshift-pipelines wait --for=condition=ready --timeout=300s pod -l app.kubernetes.io/part-of=tekton-chains
+        # ImageStreamTag to push to
+        oc -n benchmark create imagestream test
+        # SA to talk to internal registry
+        oc -n benchmark create serviceaccount perf-test-registry-sa
+        oc policy add-role-to-user registry-viewer system:serviceaccount:benchmark:perf-test-registry-sa   # pull
+        oc policy add-role-to-user registry-editor system:serviceaccount:benchmark:perf-test-registry-sa   # push
+        # Customized PipelineRun
+        dockerconfig_secret_name=$( oc -n benchmark get serviceaccount perf-test-registry-sa -o json | jq --raw-output '.imagePullSecrets[0].name' )
+        cat ./push-fake-image/run.yaml | sed "s/DOCKERCONFIG_SECRET_NAME/$dockerconfig_secret_name/g" >$TEST_RUN
+    else
+        debug "Skipping initial config as repo was already cloned"
+    fi
     # Signatures monitoring
     ./push-fake-image/measure-signed.py --server "$( oc whoami --show-server )" --namespace benchmark --token "$( oc whoami -t )" --insecure --save ./measure-signed.csv &
     measure_signed_pid=$!
